update tdarr-noding/src/tdarr/__init__.py, commands.py and logic.py
