+++ tdarr-noding/src/tdarr/__init__.py
@@ -1,2 +1,2 @@
-from .logic import Tdarr_Logic
 from .commands import Tdarr_Orders
+from .logic import Tdarr_Logic
