+++ tdarr-noding/src/tdarr/commands.py
@@ -1,5 +1,4 @@
 import requests
-from . import Tdarr_Logic
 
 
 class Tdarr_Orders:
@@ -51,40 +50,9 @@ class Tdarr_Orders:
             # pprint(response)
 
     @staticmethod
-    def set_worker_level(Server,NodeClass,set_to_level,workerType):
-        """
-        set_worker_level actually sets the worker level of a node to desired amount
-
-        Args:
-            Server (class): Server Class with endpoints
-            NodeClass (class): individual node class
-            set_to_level (int): the numeric desired to be achieved by this function
-            workerType (string): type of worker to adjust in format of transcode/healthcheck cpu/gpu all in one word
-        < Document Guardian | Protect >
-        """
-        #get node info
-        node_id=NodeClass.id_string
-
-
-        ##set worker type
-        if workerType =="All":
-            list_of_worker_types=["healthcheckcpu","healthcheckgpu","transcodecpu","transcodegpu"]
-            direction=Tdarr_Logic.get_direction(set_to_level,workerType,list_of_worker_types)
-        else:
-            direction=Tdarr_Logic.get_direction(set_to_level,workerType)
-
-        if direction != "Hold":
-            if workerType=="All":
-                list_of_payload,headers=Tdarr_Logic.payload_and_headers_worker_modification(node_id,direction,list_of_worker_types)
-            else:
-                payload,headers=Tdarr_Logic.payload_and_headers_worker_modification(node_id,direction,workerType)
-
-        if workerType=="All":
-            for payload in list_of_payload:
-                response = requests.post(
+    def mod_worker_limit(Server, headers, payload):
+        response = requests.post(
                 Server.mod_worker_limit, json=payload, headers=headers, timeout=1.5
                 )
-        else:
-            response = requests.post(
-            Server.mod_worker_limit, json=payload, headers=headers, timeout=1.5
-            )
+
+        return response
