+++ tdarr-noding/src/tdarr/logic.py
@@ -201,7 +201,7 @@ class Tdarr_Logic:
         for name in node_dictionary:
             NodeClass=node_dictionary[name]
             if NodeClass.online:
-                Tdarr_Orders.set_worker_level(Server,NodeClass,0,"All")
+                Tdarr_Logic.set_worker_level(Server,NodeClass,0,"All")
 
     ##discover up or down by set to level
     @staticmethod
@@ -263,3 +263,38 @@ class Tdarr_Logic:
                     increase_or_decrease=direction
 
         return increase_or_decrease
+
+    @staticmethod
+    def set_worker_level(Server,NodeClass,set_to_level,workerType):
+        """
+        set_worker_level actually sets the worker level of a node to desired amount
+
+        Args:
+            Server (class): Server Class with endpoints
+            NodeClass (class): individual node class
+            set_to_level (int): the numeric desired to be achieved by this function
+            workerType (string): type of worker to adjust in format of transcode/healthcheck cpu/gpu all in one word
+        < Document Guardian | Protect >
+        """
+        #get node info
+        node_id=NodeClass.id_string
+
+
+        ##set worker type
+        if workerType =="All":
+            list_of_worker_types=["healthcheckcpu","healthcheckgpu","transcodecpu","transcodegpu"]
+            direction=Tdarr_Logic.get_direction(set_to_level,workerType,list_of_worker_types)
+        else:
+            direction=Tdarr_Logic.get_direction(set_to_level,workerType)
+
+        if direction != "Hold":
+            if workerType=="All":
+                list_of_payload,headers=Tdarr_Logic.payload_and_headers_worker_modification(node_id,direction,list_of_worker_types)
+            else:
+                payload,headers=Tdarr_Logic.payload_and_headers_worker_modification(node_id,direction,workerType)
+
+        if workerType=="All":
+            for payload in list_of_payload:
+                response=Tdarr_Orders.mod_worker_limit(Server, headers, payload)
+        else:
+            response=Tdarr_Orders.mod_worker_limit(Server, headers, payload)
